docs: Add feature parity table (incomplete)
Create "prior art" table so we can see feature parity and if there are
features we're missing.

The grepcidr and perl-ipgrep tools still need investigation.

Note that grepcidr looks somewhat promising, in that it's already
available on Ubuntu. But it also appears to be slower::

    $ sudo find /etc/ -type f | sudo xargs time grepcidr 0.0.0.0/0 > grepcidr.txt
    0.09user 0.15system 0:00.25elapsed 100%CPU (0avgtext+0avgdata 17808maxresident)k
    0inputs+281outputs (0major+61796minor)pagefaults 0swaps

    $ sudo find /etc/ -type f | sudo xargs time ipgrep 0.0.0.0/0 -m w > ipgrep.txt
    0.07user 0.06system 0:00.13elapsed 99%CPU (0avgtext+0avgdata 1720maxresident)k
    0inputs+281outputs (0major+9852minor)pagefaults 0swaps

Results are the same, apart from grepcidr that matches
"::ffff:169.254.0.0/112" in 0.0.0.0/0, while ipgrep doesn't.